7.12

+++ dadp-model/src/main/java/com/gientech/pcm/depCurr/PcmDepCurr.java
@@ -1,6 +1,7 @@
 package com.gientech.pcm.depCurr;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
@@ -42,7 +43,7 @@ public class PcmDepCurr implements Serializable {
     private String currNo; // 币种
 
     @TableField(value = "BAL")
-    private Float bal; // 上日余额
+    private BigDecimal bal; // 上日余额
 
     @TableField(value = "ACCT_STS")
     private String acctSts; // 账户状态
@@ -60,7 +61,7 @@ public class PcmDepCurr implements Serializable {
     private String subAcctNo; // 活期子账户
 
     @TableField(value = "CURRENT_BAL")
-    private Float currentBal; // 当前余额
+    private BigDecimal currentBal; // 当前余额
 
     // Getters and setters
 }
