+++ dadp-model/src/main/java/com/gientech/pcm/depCurr/PcmDepCurrDTO4List.java
@@ -1,6 +1,7 @@
 package com.gientech.pcm.depCurr;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
@@ -71,7 +72,7 @@ public class PcmDepCurrDTO4List implements Serializable {
     private String currNo; // 币种
 
     @ApiModelProperty(value = "上日余额", position = 22)
-    private Double bal; // 上日余额
+    private BigDecimal bal; // 上日余额
 
     @ApiModelProperty(value = "账户状态", position = 23)
     private String acctSts; // 账户状态
@@ -89,7 +90,7 @@ public class PcmDepCurrDTO4List implements Serializable {
     private String subAcctNo; // 活期子账户
 
     @ApiModelProperty(value = "当前余额", position = 28)
-    private Double currentBal; // 当前余额
+    private BigDecimal currentBal; // 当前余额
 
     // -----------------分割线---------------------------------------
